+++ packages/terra-data-grid/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Added
   * Added screen reader messages for changes in column sort direction.
+  
+* Fixed
+  * Fixed the screen reader messages to count header row as the first row.
 
 ## 0.6.0 - (September 14, 2023)